requirements.txt
langchain
guardrails-ai
pydantic
typer[all]
python-dotenv
openai
gradio
faiss-cpu
langchain-community
langdetect
googletrans
youtube-transcript-api
watchdog
